% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibdat.R
\name{calc_calib_blr}
\alias{calc_calib_blr}
\title{Create data for calibration curves using a binary logistic regression framework with inverse probability of censoring weights}
\usage{
calc_calib_blr(
  data.mstate,
  data.raw,
  j,
  s,
  t.eval,
  tp.pred,
  curve.type = "rcs",
  rcs.nk = 3,
  loess.span,
  loess.degree,
  weights = NULL,
  w.covs,
  w.landmark.type = "state",
  w.max = 10,
  w.stabilised = FALSE,
  CI = FALSE,
  CI.R.boot,
  data.pred.plot = NULL,
  transitions.out = NULL
)
}
\arguments{
\item{data.mstate}{Validation data in msdata format.}

\item{data.raw}{Validation data in data.frame (one row per individual).}

\item{j}{Landmark state at which predictions were made}

\item{s}{Landmark time at which predictions were made}

\item{t.eval}{Follow up time at which calibration is to be assessed}

\item{tp.pred}{Vector of predicted transition probabilities at time t.eval}

\item{curve.type}{Whether calibration curves are estimated using restricted cubic splines ('rcs') or loess smoothers ('loess')}

\item{rcs.nk}{Number of knots when curves are estimated using restricted cubic splines}

\item{loess.span}{Span when curves are estimated using loess smoothers}

\item{loess.degree}{Degree when curves are estimated using loess smoothers}

\item{weights}{Vector of inverse probability of censoring weights}

\item{w.covs}{Character vector of variable names to adjust for when calculating inverse probability of censoring weights}

\item{w.landmark.type}{Whether weights are estimated in all individuals uncensored at time s ('all') or only in individuals uncensored and in state j at time s ('state')}

\item{w.max}{Maximum bound forinverse probability of censoring weights}

\item{w.stabilised}{Indicates whether inverse probability of censoring weights should be stabilised or not}

\item{CI}{Size of confidence intervals as a \%}

\item{CI.R.boot}{Number of bootstrap replicates when estimating the confidence interval for the calibration curve}

\item{data.pred.plot}{Dataframe or matrix of predicted risks for each possible transition over which to plot the calibration curves. Must have one column for every possible transition.}

\item{transitions.out}{Transitions for which to calculate calibration curves. Will do all possible transitions if left as NULL.}
}
\description{
Create data for calibration curves using a binary logistic regression framework with inverse probability of censoring weights
}
