% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibdat.R
\name{calc_calib_mlr}
\alias{calc_calib_mlr}
\title{Create data for calibration curves using a binary logistic regression framework with inverse probability of censoring weights}
\usage{
calc_calib_mlr(
  data.mstate,
  data.raw,
  j,
  s,
  t.eval,
  tp.pred,
  smoother.type,
  ps.int = 4,
  degree = 3,
  s.df = 4,
  niknots = 4,
  weights = NULL,
  w.covs,
  w.landmark.type = "state",
  w.max = 10,
  w.stabilised = FALSE
)
}
\arguments{
\item{data.mstate}{Validation data in msdata format.}

\item{data.raw}{Validation data in data.frame (one row per individual).}

\item{j}{Landmark state at which predictions were made}

\item{s}{Landmark time at which predictions were made}

\item{t.eval}{Follow up time at which calibration is to be assessed}

\item{tp.pred}{Vector of predicted transition probabilities at time t.eval}

\item{smoother.type}{s, sm.os or sm.ps}

\item{ps.int}{the number of equally-spaced B spline intervals in the vector spline smoother (see VGAM::sm.ps)}

\item{degree}{the degree of B-spline basis in the vector spline smoother (see VGAM::sm.ps)}

\item{s.df}{degrees of freedom of vector spline (see VGAM::s)}

\item{niknots}{number of interior knots for VGAM::sm.os}

\item{weights}{Vector of inverse probability of censoring weights}

\item{w.covs}{Character vector of variable names to adjust for when calculating inverse probability of censoring weights}

\item{w.landmark.type}{Whether weights are estimated in all individuals uncensored at time s ('all') or only in individuals uncensored and in state j at time s ('state')}

\item{w.max}{Maximum bound forinverse probability of censoring weights}

\item{w.stabilised}{Indicates whether inverse probability of censoring weights should be stabilised or not}
}
\description{
Create data for calibration curves using a binary logistic regression framework with inverse probability of censoring weights
}
