% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{calc_weights_OLD}
\alias{calc_weights_OLD}
\title{Calculate inverse probability of censoring weights for a group of individuals in state j at time s}
\usage{
calc_weights_OLD(
  data.mstate,
  data.raw,
  covs,
  t.eval,
  s,
  landmark.type = NULL,
  j = NULL,
  max.weight = 10,
  stabilised = FALSE
)
}
\arguments{
\item{data.mstate}{Validation data in msdata format}

\item{data.raw}{Validation data in data.frame (one row per individual)}

\item{covs}{Character vector of variable names to adjust for when calculating inverse probability of censoring weights}

\item{t.eval}{Follow up time at which to calculate weights}

\item{s}{Landmark time at which predictions were made}

\item{landmark.type}{Whether weights are estimated in all individuals uncensored at time s ('all') or only in individuals uncensored and in state j at time s ('state')}

\item{j}{Landmark state at which predictions were made (only required in landmark.type = 'state')}

\item{max.weight}{Maximum bound for weights}

\item{stabilised}{Indicates whether weights should be stabilised or not}
}
\description{
Calculate inverse probability of censoring weights for a group of individuals in state j at time s
}
