% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibmsm-package.R
\docType{package}
\name{calibmsm-package}
\alias{calibmsm}
\alias{calibmsm-package}
\title{calibmsm: Calibration plots for the transition probabilities from a multistate model}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Alexander Pate \email{alexander.pate@manchester.ac.uk} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
