% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ebmtcal}
\alias{ebmtcal}
\title{European Group for Blood and Marrow Transplantation data}
\format{
\subsection{'ebmtcal'}{

A data frame with 2,279 rows and 17 columns:
\describe{
\item{id}{Patient indentifier}
\item{rec, rec.s}{Time until and event indicator for recovery variable}
\item{ae, ae.s}{Time until and event indicator for adverse event variable}
\item{recae, recae.s}{Time until and event indicator for recovery + adverse event variable}
\item{rel, rel.s}{Time until and event indicator for relapse variable}
\item{srv, srv.s}{Time until and event indicator for death variable}
\item{year}{Year of transplant}
\item{agecl}{Age at transplant}
\item{proph}{Prophylaxis}
\item{match}{Donor-recipient match}
\item{dtcens}{Time of censoring}
\item{dtcens.s}{Event indicator, 1:censoring occured, 0: absorbing state entered before censoring occured}
}
}
}
\source{
\url{https://search.r-project.org/CRAN/refmans/mstate/html/EBMT-data.htmlE}
}
\usage{
ebmtcal
}
\description{
ebmt data in format ready to assess calibration
}
\keyword{datasets}
