% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibplot.R
\name{plot.calib_mlr}
\alias{plot.calib_mlr}
\title{Create data for calibration curves using a binary logistic regression framework with inverse probability of censoring weights}
\usage{
\method{plot}{calib_mlr}(x, ..., combine = TRUE, ncol = NULL, nrow = NULL)
}
\arguments{
\item{x}{Object of type 'calib_mlr' generated from calc_calib_mlr}

\item{...}{Other}

\item{combine}{Whether to combine into one plot using ggarrange, or return as a list of individual plots}

\item{ncol}{Number of columns for combined calibration plot}

\item{nrow}{Number of rows for combined calibration plot}
}
\description{
Create data for calibration curves using a binary logistic regression framework with inverse probability of censoring weights
}
