% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tps0}
\alias{tps0}
\title{Predicted transition probabilities out of every state at time s = 0}
\format{
\subsection{'tps0'}{

A data frame with 13,674 rows and 14 columns:
\describe{
\item{id}{Patient indentifier}
\item{pstate1, pstate2, pstate3, pstate4, pstate5, pstate6}{Predicted transition probabilities of transitions into states 1 to 6}
\item{se1, se2, se3, se4, se5, se6}{Standard error of the predicted transition probabilities of transitions into states 1 to 6}
\item{j}{State from which the predicted transition probabilities are estimated from}
}
}
}
\usage{
tps0
}
\description{
Predicted transition probabilities out of every state at time s = 0
}
\keyword{datasets}
