% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibmsm-package.R
\docType{package}
\name{calibmsm-package}
\alias{calibmsm}
\alias{calibmsm-package}
\title{calibmsm: Calibration plots for the transition probabilities from multistate models}
\description{
A toolbox to aid in the estimation of calibration plots for multistate models which have been developed. Calibration of the transition probabilities at time t when in state j at time s are assessed. This is done using a combination of landmarking, calibration methods for binary logistic and multinomaial logistic regression models and inverse probability of censoring weights.
}
\author{
\strong{Maintainer}: Alexander Pate \email{alexander.pate@manchester.ac.uk} (\href{https://orcid.org/0000-0002-0849-3458}{ORCID})

}
\keyword{internal}
