% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibplot.R
\name{plot.calib_blr}
\alias{plot.calib_blr}
\title{Plot calibration curves}
\usage{
\method{plot}{calib_blr}(x, ..., combine = TRUE, ncol = NULL, nrow = NULL)
}
\arguments{
\item{x}{Object of type 'calib_blr' generated from calc_calib_blr}

\item{...}{Other}

\item{combine}{Whether to combine into one plot using ggarrange, or return as a list of individual plots}

\item{ncol}{Number of columns for combined calibration plot}

\item{nrow}{Number of rows for combined calibration plot}
}
\description{
Creates calibration plots for the transition probabilities of a multistate model
using data estimated from \code{calc_calib_blr}. Such data was estimated using the
binary logistic regression framework with inverse probability of censoring weights.
}
